seprates CKDTableInfo from api to contracts and uses it in exomolop.py
