+++ src/exojax/opacity/ckd/api.py
@@ -7,7 +7,6 @@ maintaining accuracy through k-distribution statistical representation.
 
 from __future__ import annotations
 from typing import Union, Optional
-from dataclasses import dataclass
 import json
 
 import jax.numpy as jnp
@@ -15,37 +14,16 @@ import numpy as np
 from jax import vmap
 
 from exojax.opacity.base import OpaCalc
+from exojax.opacity.ckd.contracts import CKDTableInfo
 from exojax.opacity.ckd.core import gauss_legendre_grid
 from exojax.opacity.ckd.core import compute_ckd_tables
 from exojax.opacity.ckd.core import interpolate_log_k_2d
 from exojax.opacity.ckd.io import _hash_json
 from exojax.opacity.ckd.io import _base_fingerprint
 from exojax.opacity.ckd.io import _ckd_save_as_npz
-
 from exojax.utils.spectral_bands import spectral_bands
 
 
-@dataclass(frozen=True)
-class CKDTableInfo:
-    """Immutable container for CKD table information.
-
-    Attributes:
-        log_kggrid: Log k-values on g-grid, shape (nT, nP, Ng, nnu_bands)
-        ggrid: Gauss-Legendre g-ordinates, shape (Ng,)
-        weights: Gauss-Legendre quadrature weights, shape (Ng,)
-        T_grid: Temperature grid, shape (nT,)
-        P_grid: Pressure grid, shape (nP,)
-        nu_bands: Wavenumber band centers, shape (nnu_bands,)
-        band_edges: Wavenumber band edges, shape (nnu_bands, 2)
-    """
-
-    log_kggrid: jnp.ndarray
-    ggrid: jnp.ndarray
-    weights: jnp.ndarray
-    T_grid: jnp.ndarray
-    P_grid: jnp.ndarray
-    nu_bands: jnp.ndarray
-    band_edges: jnp.ndarray
 
 
 class OpaCKD(OpaCalc):
